Fixed algorithmic error in ringbuffer.h This caused the ringbuffer to be unable to not report correct values when rb_isFull() is run.
Also fixed error in usart code where UDRE interrupt was
disabled/enabled where it should have been the interrupt flag
that should be raised/lowered.

The board is no longer crashing when run with buffered output
but it does not behave as expected, therefor issue #2 still applies.
When testing with the io_read_write test node that has a soft delay
at start of it's mainloop it appears as if this delay is prolonged
when using buffered output, as in it prints much slower than the
delay specifies.