Fixed bug that caused paddleshift interrupt to fire without being triggered.
The root cause of this was that the paddleshift interrupt was sat to be
asynchronous interrupts. Setting them to synchronous interrupts fixed
the problem.

With this bug fixed we are no longer delated by constantly executing the
isr so what looked like cpu clock problems are also solved. Therefor we
have changed some delays from us delays to ms delays