Moved ecu package names into a lookup macro
This removes some memory overhead of redundant packages.

Also changed sensor value to be a float instead of a double. The reason
for this is twofold. Firstly, we get a 32bit raw value from the ecu so
the max resolution is equal to that of a float anyway. Also when
compiling with avr-gcc sizeof(double)==sizeof(float)==4 anyway