New "simplified" CAN driver. A new CAN driver for sending and receiving CAN messages not limited to 8 bytes length.
This new driver is a lot more complicated
in itself but a whole lot easier to use.
It makes use of the newly implemented heap
to store messages during sending and receiving
and automaticly frees up memory when finnished.

The example can_simple_rx_tx shows how to use
the driver to send longer CAN messages with a
simple function taking three parameters
(id, message length and the message itself).

The driver aims to comply with the standard
ISO 15765-2. As of now only a subset of the
standard has been implemented, just enough
for it to be usable.
A description of the standard can be found at
https://en.wikipedia.org/wiki/ISO_15765-2