Modified all io calls to use the stdio functions.
Note that this may have yet undiscovered side effects that breaks stuff.

fx. Because we now use buffered io in the ComNode are the xbee no longer
capable of handeling how fast we push data into it causing it to drop
packages. This will be fixed by lowering the baudrate so we cannot push
data faster than it can handle