event_manager: Renamed event_manager driver to cpu_load
The event manager in the libat90 drivers, has been changed so
that it no longer handles any events.
I figured that in the end there was no reason/need for this
functionality and that the only interesting task to leave in
would be a counter to track the cpu load.
Hence the name change.

Due to the change it has been removed from nodes where it was used.
Since most of the nodes are no longer used, the cpu_load counter
should be reimplemented where needed.