libat90: added simple untestet interface to convert half-floats.
Half-floats could in the future be used to compress data
for logging.
This implementation however is a super simple and fast
conversion whith shortcommings. This is not likely to
be an issue with the data we recieve but should nontheless
be worked on in the future for more complete support.

Shortcommings include non-support for infinite and NAN.