libat90/system_messages: Changed the way we list messages.
In the past it has been an annoyance that we linked from
system_messages in libat90 back to a header in the ComNode,
because we needed an enum of the ECU packages.
Since the system_messages module depends on knowing the enum
at compile time,
we would be better suited with finding a better way to share
this information, or the system_messages module.
It's a relative large change this late in the development,
so until elsewhere has been agreed upon I have moved the content
of the enums into some seperate lists in libat90/lists.
this makes for a more dynamic inclusion into the code,
and is a somewhat nicer solution and a brigde to implementing
a more permanent solution in the future.