libat90/system_messages: Changed the transport method to be set at runtime.
Before the way a messages would have to be transmitted was set at compile-time
which meant a very static and inflexible communications system.
It also meant that we would have to change in the libat90 code every time
we wanted a message to be handled differently by one of the nodes.
Lastly it meant that every node would have to threat messages the same
which is simply an unnecessary restriction.

The changes adds new functions to this module to change the method of transportation.
This means that every node from now on have to specify what is to happen with
collected information in the same way that they have to subscribe for
incomming messages.