ComNode/send_file: A dedicated file for request file and handles both ACK and NACK.
The request REQUEST_FILE has been been given a few functions to handle
different cases. For this amount of code it has been given its own file.
Both ACK and NACK are now handled (RESEND not handled correctly at this stage),
and can easily be expanded to be used for other requests as well.

A better way to handle errors and request status is needed.