ComNode/protocol: Made a seperate function for setting ongoing_request.
Made a function which can be called outside protocol.c to set
the ongoing_request variable.
This makes it a little easier to keep track on what it is doing
without a bunch of return values all over the place.
Hopefully also fixes a bug in the send file request,
which let to too many bytes being transmitted.