ComNode/protocol: Fixed that ECU stops anwering completely after stop button is pushed.
We had the problem that whenever the STOP button was pushed,
which it often is when racing, we would stop getting data from
the ECU even after the button was released again.

The issue here is that when the button is pushed, power to
the engine and the ECU is cut, but not to the Nodes.
So when the button is released again the ComNode does not know
that power has been cut and that it is supposed to send a new
data request to the ECU.

This fix checks if we have gotton a full data packet from the ECU
in the last 300 ms. If that is not the case the ECU input ringbuffer
is emptied for whatever data would be left, and a new request is
send to the ECU.